requirements.txt
python-telegram-bot==20.7
requests
python-dotenv
redis
nest_asyncio
apscheduler
